requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-debug-toolbar==2.1
django-webpack-loader==0.6.0
djangorestframework==3.11.0
idna==2.8
lxml==4.4.2
Pillow==7.0.0
pytz==2019.3
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
